+++ astropy/io/ascii/basic.py
@@ -155,10 +155,13 @@ class CommentedHeader(Basic):
         """
         out = super().read(table)
 
-        # Strip off first comment since this is the header line for
-        # commented_header format.
+        # Strip off the comment line set as the header line for
+        # commented_header format (first by default).
         if 'comments' in out.meta:
-            out.meta['comments'] = out.meta['comments'][1:]
+            idx = self.header.start_line
+            if idx < 0:
+                idx = len(out.meta['comments']) + idx
+            out.meta['comments'] = out.meta['comments'][:idx] + out.meta['comments'][idx+1:]
             if not out.meta['comments']:
                 del out.meta['comments']