+++ CHANGES.rst
@@ -668,6 +668,9 @@ astropy.extern
 astropy.io.ascii
 ^^^^^^^^^^^^^^^^
 
+- Fix stripping correct (header) comment line from ``meta['comments']``
+  in the ``CommentedHeader`` reader for all ``header_start`` settings. [#7508]
+
 astropy.io.fits
 ^^^^^^^^^^^^^^^